Renamed initializer with an underscore
